Remove dagger + clean up dependencies
