+++ app-example/build.gradle
@@ -2,7 +2,6 @@ configureAndroidApp(project)
 
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
-apply plugin: 'kotlin-kapt'
 
 def getAccessToken() {
     def systemEnvKey = System.getenv('ACCESS_KEY')
@@ -57,16 +56,10 @@ android {
 dependencies {
     implementation deps.external.kotlinStdlib
 
-    kapt deps.apt.daggerCompiler
-
     implementation deps.android.appCompat
     implementation deps.android.constraintLayout
     implementation deps.android.design
 
-    implementation deps.external.dagger
-    implementation deps.external.mviCore
-    implementation deps.external.mviCoreAndroid
-    implementation deps.external.mviCoreDiff
     implementation deps.external.rxrelay2
     implementation deps.external.rxandroid2
 
@@ -76,15 +69,14 @@ dependencies {
     implementation deps.retrofit.okhttp
     implementation deps.retrofit.okhttpLoggingInterceptor
     implementation deps.retrofit.picasso
-    implementation project(":libraries:rib-base")
 
     implementation project(":libraries:rib-base")
     implementation project(":libraries:rib-portal")
 
     implementation deps.external.mviCore
     implementation deps.external.mviCoreAndroid
+    implementation deps.external.mviCoreDiff
 
-    testImplementation deps.external.roboelectricBase
     testImplementation deps.test.junit
     testImplementation deps.test.mockitoKotlin
     testImplementation deps.test.assertj
