+++ app-example/src/main/java/com/badoo/ribs/example/app_bar/AppBarInteractor.kt
@@ -1,27 +1,22 @@
-package com.badoo.ribs.example.component.app_bar
+package com.badoo.ribs.example.app_bar
 
 import androidx.lifecycle.Lifecycle
 import com.badoo.mvicore.android.lifecycle.startStop
 import com.badoo.mvicore.binder.using
 import com.badoo.ribs.clienthelper.interactor.Interactor
 import com.badoo.ribs.core.modality.BuildParams
-import com.badoo.ribs.example.component.app_bar.analytics.AppBarAnalytics
-import com.badoo.ribs.example.component.app_bar.builder.AppBarBuilder.Params
-import com.badoo.ribs.example.component.app_bar.mapper.UserToViewModel
-import com.badoo.ribs.example.component.app_bar.mapper.ViewEventToAnalyticsEvent
-import com.badoo.ribs.example.component.app_bar.mapper.ViewEventToOutput
-import com.badoo.ribs.example.component.app_bar.routing.AppBarRouter.Configuration
+import com.badoo.ribs.example.app_bar.AppBarBuilder.Params
+import com.badoo.ribs.example.app_bar.mapper.UserToViewModel
+import com.badoo.ribs.example.app_bar.mapper.ViewEventToOutput
 import com.badoo.ribs.example.network.model.User
 import com.badoo.ribs.example.repository.UserRepository
-import com.badoo.ribs.routing.source.backstack.BackStackFeature
 import io.reactivex.ObservableSource
 import io.reactivex.functions.Consumer
 
 internal class AppBarInteractor(
     buildParams: BuildParams<Params>,
     userRepository: UserRepository,
-    private val output: Consumer<AppBar.Output>,
-    private val backStack: BackStackFeature<Configuration>
+    private val output: Consumer<AppBar.Output>
 ) : Interactor<AppBar, AppBarView>(
     buildParams = buildParams
 ) {
@@ -34,7 +29,6 @@ internal class AppBarInteractor(
     override fun onViewCreated(view: AppBarView, viewLifecycle: Lifecycle) {
         viewLifecycle.startStop {
             bind(view to output using ViewEventToOutput)
-            bind(view to AppBarAnalytics using ViewEventToAnalyticsEvent)
             bind(userSource to view using UserToViewModel)
         }
     }
