+++ app-example/src/main/java/com/badoo/ribs/example/app_bar/AppBarNode.kt
@@ -1,0 +1,16 @@
+package com.badoo.ribs.example.app_bar
+
+import android.view.ViewGroup
+import com.badoo.ribs.core.Node
+import com.badoo.ribs.core.modality.BuildParams
+import com.badoo.ribs.core.plugin.Plugin
+
+class AppBarNode internal constructor(
+    buildParams: BuildParams<*>,
+    viewFactory: ((ViewGroup) -> AppBarView?)?,
+    plugins: List<Plugin> = emptyList()
+) : Node<AppBarView>(
+    buildParams = buildParams,
+    viewFactory = viewFactory,
+    plugins = plugins
+), AppBar
