+++ app-example/src/main/java/com/badoo/ribs/example/app_bar/mapper/ViewEventToOutput.kt
@@ -1,7 +1,7 @@
-package com.badoo.ribs.example.component.app_bar.mapper
+package com.badoo.ribs.example.app_bar.mapper
 
-import com.badoo.ribs.example.component.app_bar.AppBar.Output
-import com.badoo.ribs.example.component.app_bar.AppBarView.Event
+import com.badoo.ribs.example.app_bar.AppBar.Output
+import com.badoo.ribs.example.app_bar.AppBarView.Event
 
 internal object ViewEventToOutput : (Event) -> Output? {
