--- app-example/src/main/java/com/badoo/ribs/example/component/app_bar/builder/AppBarScope.kt
@@ -1,7 +1,0 @@
-package com.badoo.ribs.example.component.app_bar.builder
-
-import javax.inject.Scope
-
-@Scope
-@Retention(AnnotationRetention.RUNTIME)
-internal annotation class AppBarScope
