--- app-example/src/main/java/com/badoo/ribs/example/component/app_bar/routing/AppBarChildBuilders.kt
@@ -1,7 +1,0 @@
-package com.badoo.ribs.example.component.app_bar.routing
-
-internal class AppBarChildBuilders
-// TODO add builders through constructor if built from dagger module:
-// (
-    //val child1: Child1Builder
-//)
